requirements.txt
python-dotenv==1.2.1 
pyyaml==6.0.3
requests==2.32.5
playwright==1.57.0